CIS 1.2.21: Ensure that the metrics are protected by RBAC
Vanilla CIS k8s benchmark prescribes that profiling should be disabled.
In OCP however, profiling is on by default and used by Prometheus.

So instead of making sure profiling is disabled, let's make sure it is
explicitly protected by RBAC by checking the cluster-debugger role for
including the /metrics endpoint

The previous rule is unused now and was removed.