Reorganize the env_yaml and product_yaml
Separate into products.py functions that are related to products,
including checking its data and deriving properties.
Relocate open_environment() away from yaml.py, this function is more
about loading build environment configurations than yaml loading per se.

This should ease the life of scripts to only want to load data from the
product, but don't care about the environment.
(How dare we have scripts that are not eco friendly?)

Note: this commit breaks the build as other scripts need to be updated.
It is done as a separate commit to help understand the changes.