rhcos4: Add check for nousb kernel argument
This adds a check that verifies if the nousb kernel argument is set. If
it isn't it should generate an appropraite remediation.

This also enhances the `coreos_kernel_option` template and the
`oval_argument_value_in_line_criterion` and
`oval_argument_value_in_line_test` macros to not require `arg_value`
anymore. If this parameter is not given, it'll only use `arg_name`.

Signed-off-by: Juan Antonio Osorio Robles <jaosorior@redhat.com>