RHCOS4: Remove rules that use rpmverifypackage_test
These are meant to check that files match the ownership and permissions
that the RPMS define... however, in RHCOS these checks are not reliable
as ostree might handle some of these files differently.

On the other hand, ostree does ensure that some of these are kept in a
correct state as it mounts /usr as read-only. We'll ultimately address
the integrity of files using FIO/AIDE, so we don't completely loose
compliance here.

Signed-off-by: Juan Antonio Osorio Robles <jaosorior@redhat.com>