Updates to support ansible_playbook_to_role utility
Numerous updates were made to enhance the functionality
and to fix bugs:
* Updated GH actions to run on PR's. Also added new tag skip.
* Updated galaxy meta template. Includes additions of variables
to be used as substitutions to support automating the creation
of that file, and to prevent manual additions as was handled in
the past.
* Added the anssi profiles to the whitelist (have not done any testing
on these roles yet)
* Updates/bug fixes to support meta templating. Fixed issue with
meta and readme files being overwritten with the wrong output.
* Updated syntax for python3 support. No need to use python2 for
this IMO.