Create CPE per OCP4 version to handle deprecations
Since OCP4 uses the same datastream for multiple version, the time has
come for us to start differentiating between what's supported in each
version. Thus, we have now specific CPEs for different versions that can
be used to delimit what's supported.

A first instance of this kind of problem is the flowschema check:

It had been in v1alpha1 for a while, when it was promoted to v1beta1 in
OpenShift 4.8 (Kubernetes version 1.21). Thus, we promoted the rule that
was previously checking this to use the newer version, and created a new
rule for the old version.

The each of these rules now has the appropriate supported versions
marked on the `platforms` key.

Signed-off-by: Juan Antonio Osorio Robles <jaosorior@redhat.com>