Prevent unqualified CIS and STIGID references
Per discussion on #6416, I don't think it is generally valid to have a
unqualified (by product) CIS reference identifier. While there is a
hierarchy to the CIS benchmarks (from General Linux to e.g., Debian
Family to e.g., Ubuntu), references aren't necessarily the same across
families (e.g., Ubuntu to RHEL) and order of the sections can differ
significantly. The same holds for STIG: unless a product undergoes STIG
certification (thus assigning rules a specific STIG ID for the product),
we shouldn't reuse one product's STIG IDs for other, unrelated products.

In the future, we might want to offer a "hybrid" form (e.g., cis@ubuntu)
applying to all versions of a product, but otherwise I think this is the
best route to go.

Start enforcing this at build time, to ensure we (and other
contributors) don't accidentally undo this work in the future.

Signed-off-by: Alexander Scheel <alex.scheel@canonical.com>