Add --dry-run option to fix_rules.py
This tells fix_rules to run without writing any changes out to disk and
without prompting the caller as to whether or not to write out changes.
It additionally suppresses the comparison output and only shows which
files need changing.

The return status has also changed. Previously this utility's exit
status was always 0. Now it will return 1 if any changes need to be made
(or have been made as a result of this program's execution) or 0 if no
changes need be made.

Signed-off-by: Alexander Scheel <alexander.m.scheel@gmail.com>