Check all rules in profile for required reference
The new refchecker.py utility automatically checks all rules in a given
profile for a required reference field. This is most commonly useful
when creating benchmarks for a specified system, such as CIS or STIG,
where it is expected that all rules in a given benchmark have references
related to them.

Unlike tests/missing_refs.sh (and its helper tool,
build-scripts/profile_tool.py), this operates on the YAML level and
should have fewer false-positives in the content for non-RHEL
identifiers. Additionally, this parses the present values of the rules,
meaning that it can be used between builds.

Signed-off-by: Alexander Scheel <alex.scheel@canonical.com>