ocp4: Enhance group ownership checks openvswitch processes pid files
In a more recent version of Open vSwitch, the process pid files changed
ownership from "openvswitch" to "hugetlbfs". OCP 4.6 will stay in an
older version of OvS, so to keep this working with newer versions (e.g.
the one included in OCP 4.8), this fixes the content to check for both
groups. This way the content keeps being valid accross releases.

Closes https://github.com/ComplianceAsCode/content/issues/6913

Signed-off-by: Juan Antonio Osorio Robles <jaosorior@redhat.com>