Remove support statement from the manual page
This statement is outdated. It says that SSG is shipped in EPEL and that
it isn't supported by Red Hat. We don't ship it EPEL for many years.
When the man page is shipped in RHEL it isn't true that it isn't
supported by Red Hat. When the man page is packaged in other
distributions, the information about support by Red Hat is redundant.
The package usually follows the distro's support policy.  Other packages
also don't claim their support level in their manual page. People should
refer to the distribution support documentation instead.