Add option_regex_suffix to sudo_defaults_option template
In the course of porting Ubuntu CIS tooling to CaC, we too used the
sudo_defaults_option template for the logfile rule. However, we found a
shortcoming in the existing template: `\w` matches word characters and
not `/` that many of the log file paths will use. Introduce an optional
template variable, option_regex_suffix, to control the regex portion
_after_ the option name in this template. This will allow us to specify
a different value in the sudo_custom_logfile rule.

Signed-off-by: Alexander Scheel <alex.scheel@canonical.com>