Attempt to detect issues with sed substitution
When using Jinja to perform substitution into sed expressions, we need
to be careful to avoid introducing expression separators from the
substituted variables. E.g., if the command is:

    sed 's/{{{ substitution }}}//'

We need to ensure `/` isn't present in `substitution`.

While this won't help with any XCCDF variables having bad values, this
will help with easily-detected build-time failures.

Signed-off-by: Alexander Scheel <alex.scheel@canonical.com>