Refactor ssg/templates.py generation (SCE)
This refactors SSG's template generation code to better align between
where we are now upstream and future changes related to templated SCE
support.

At this point in time, we wish to make some changes to allow building
templated (both in the Jinja and shared/templates senses) test cases,
and many of the SCE-enablement changes are relevant, I chose to pull
them in almost entirely. The original commit message is preserved below:

    Refactor template generation for SCEs

    ssg/templates.py holds the core of template generation logic. We
    will need a template_builder for SCE (so we can load SCE metadata
    from templated content) but we do not wish to write this SCE content
    out immediately; instead, we wish to wait until
    build-scripts/build_templated_content.py is run.

    We refactor to make the resolved languages for a rule (the
    intersection between theoretical languages allowed by the rule and
    the actual languages allowed by the template) accessible to callers,
    and, also allow reading just a single templated language artifact
    into memory (instead of from disk).

Notably, this last change is most useful to us; we don't want to put the
file in a template/build-system specified location; we wish to control
its location exactly here.

Note that no changes to ssg/templates.py for the test suite change were
done here; this was purely a change to sync future changes for SCE
content.

Signed-off-by: Alexander Scheel <alex.scheel@canonical.com>