Add templated tests into templated directory
This extends the existing templated directory generation to also include
tests under shared/templates/.../tests. The complicated portion is in
subdirectory handling: we need to ensure we create all necessary nested
subdirectories that are implicitly implied by the relative paths. This
is a shortcoming in the get_all_tests's return format.

Signed-off-by: Alexander Scheel <alex.scheel@canonical.com>