Generate more reproducible remediations
We should sort the given generated conditionals of the remediations
(Bash and Ansible) so that when we compare the datastreams, they're
more similar by default.

When comparing datastreams, I saw a lot of entries like:

    -if [ "$var_firewall_package" == "iptables" ] || [ ! -f /.dockerenv ] && [ ! -f /run/.containerenv ]; then
    +if [ ! -f /.dockerenv ] && [ ! -f /run/.containerenv ] || [ "$var_firewall_package" == "iptables" ]; then

These are exactly the same, just different ordering due to the set.
IMO, the set is important, so we should sort it afterwards.

This helps to reduce noise in compare-ds.py.

Signed-off-by: Alexander Scheel <alex.scheel@canonical.com>