Template tests with matching prodtype
When prodtype is known to the test system, we can skip templating any
rule that has a prodtype that doesn't match. This saves time during
building the bundle and restricts us to only writing tests we care
about and can potentially use.

Note that there might be a mismatch between (datastream, rule): if a
rule is updated on disk but the datastream isn't regenerated, you might
run into weird edge cases where this either over- or under-provisions,
but the same issue would likely occur previously.

Signed-off-by: Alexander Scheel <alex.scheel@canonical.com>