Remove FragmentPath check from service_disabled
In https://github.com/systemd/systemd/issues/582 it is documented that
systemd could eventually replace FragmentPath=/dev/null (on masked
services) with the actual service path -- not the fully (symlink)
resolved path as is currently the case.

This matches the behavior currently seen in Ubuntu (all versions) and
RHEL 9/Fedora 34.

Per discussion with Gabriel, Matej, Richard, and Matt, it is best to
remove this check, especially since ActiveState=Masked suffices.

Resolves: #7280
Resolves: #7248

Signed-off-by: Alexander Scheel <alex.scheel@canonical.com>