Fix combined mode execution in SSGTS
While enabling GH-action based execution of the SSGTS, an option for
executing multiple rules was added, allowing for comma-separated rules
to be passed via the command line. This doesn't behave well with
combined mode, which expected a single target (a string) versus a list.

Update combined mode to understand multiple arguments via the CLI and
execute each profile separately.

Signed-off-by: Alexander Scheel <alex.scheel@canonical.com>