Remove package applicability from s390x_arch generated remediations.
This makes sure that there will be no package applicability check in the
remediations of rules that use the s390x_arch platform applicability,
since the check is made by checking a line in file instead. At this
moment the build system does not allow doing such checks. The side
effect is that Bash and Ansible roles will apply this remediation even
on a system that is not s390_arch, so using OpenSCAP scanner is highly
recommended.