linux_os/guide/system/accounts/accounts-pam/locking_out_password_attempts/accounts_password_pam_unix_remember/: handle pwhistory.so
The test suite is currently failing since the rhel8_wrong_value.fail.sh is adding a new
line for pam_pwhistory.so. Then, the bash remediation script is looking for a line with
remember= in order to know if it must change or add "remember=$value". In the test suite,
it'll match the pam_pwhistory.so line so will try to change its value, but does nothing
since the sed line is looking for pam_unix.
So, fix the bash remediation script to properly look for pam_unix.so line and add support
to ansible and the bash remediation to handle pam_pwhistory.

Signed-off-by: Arnaud Patard <apatard@hupstream.com>