Implement `filter_rules` for Profiles
This implements a mechanism to filter rules for a profile based on
attributes of the rule object.

This allows us to do stuff like this:

```
filter_rules: 'platform == "ocp4-node"'
```

... Since the filter takes the Python syntax, more complex rules can be
created. e.g. one is able to use all boolean operators from the
language.

The variables available are exactly the ones available in the `Rule`
object from *build_yaml.py*.

Signed-off-by: Juan Antonio Osorio Robles <jaosorior@redhat.com>