Make `filter_rules` work for regular profiles too
This makes the filter_rules mechanism to work for profiles built
normally (without the controls structure).

This required turning the selections and unselections to be Rule objects
to enable easy filtering. Filtering is now implemented in the control
resolution (`resolve_controls`) as opposed to the `_merge_controls`
function. And the Rule object had to be enhanced to support comparison
functions for it to effectively work in a `set`.

As a side-effect, the usage of `env_yaml` for profiles is now required.
Changes had to be made to allow the tools depending on Profiles to take
this into use.

Signed-off-by: Juan Antonio Osorio Robles <jaosorior@redhat.com>