Add new template to check that Compliance Operator returned NotFound
A recent addition to the compliance operator [1] added the ability to
mark an empty file with the `NotFound` error, this allows to do
compliance checks on whether the object was found on the cluster or not.

This includes that functionality into an OVAL probe which verifies the
expected file and checks the needed string. This way, we can do a
relevant check.

[1] Commit f200ae00f32ea29670cfa67ee802109b0a0709e2

Signed-off-by: Juan Antonio Osorio Robles <jaosorior@redhat.com>