OCP: Fix OCIL description of configure_network_policies_namespaces
The JQ filter used in both the warnings for API collection purposes and
in OCIL was user-unfriendly as it was dumping the whole namespace
object. Because just changing the filter to look nicer would break the
API collection, I added a similar variable, just filtering the names
only and used that in OCIL only.

Jira: OCPBUGSM-35010
RHBZ: 2003170