Fix bash remediation spurious variable substitution
If a variable value contains the char '$', the resulting fix in the
XCCDF file will generate the following pattern:

<variable>="<value with a '$' char>"

so when the bash engine executes the code, it erroneously finds a
new variable and substitutes its value. This leads to a variable
value which is different from the value set in the XCCDF or
tailoring file.

This fix only changes the double quotes enclosing the value by single
quotes, thus preventing the above behavior.

Signed-off-by: Richard Maciel Costa <richard.maciel.costa@canonical.com>