Add ability to generate OCIL directly
This commit adds the option to build OCIL file to the build_shorthand.py
script. Currently, the OCIL data are present in compiled rules. The
build_shorthand.py script iterates over compiled rules when composing
the shorthand and puts the data to the generated shorthand XML. Then, a
series of XSLT transformations is employed during the build process to
extract the OCIL from the shorthand XML and to replace OCIL data in the
shorthand XML by a reference. Instead, we can create the OCIL file
directly, during processing of rule data because we already have the
data in the memory. That will enable us to simplify the build process in
future. We won't need to process the shorthand XML to extract the OCIL,
we will be able to use the OCIL generated by this script and therefore
we won't need some XSLT transformation modifying shorthand XML.