Properly add local_variables to cpe-oval component
Any variable referenced by a CPE OVAL check (inventory) needs to be
added to the cpe-oval component.

Tests were searched variables references, but actually the objects and
states are the elements that can reference a variable.

External variables are not useful during CPE evaluation, so the script
throws an error when one is found.