OCP4: Allow cluster filtering macros to work accross rules
The Compliance Operator has a mechanism for allowing filtering of
resources through `jq` syntax. This is quite powerful, but has some
content caveats:

In order to ensure that folks can filter the same Kube resource in
different rules, we need to ensure that they don't stumble upon each
other. If we don't, they'll overwrite the same file, and we won't get
reliable results.

Thus, we introduces a mechanism in CaC through jinja to generate a hash
ID for the filtered files.

This mechanism relied on `rule_id + path + filter`, which was passed to
a hashing function to generate the ID.

This, unfortunately, breaks down when importing a check from another
rule. Since we need these hashes to match. It resulted in the rule not
evaluating as it should (the file wasn't being found as the hash didn't
match).

This removes the `rule_id` from the function. Thus allowing us to share
these checks between rules, while still having a unique ID based on the
kube API path and the filter.

Signed-off-by: Juan Antonio Osorio Robles <jaosorior@redhat.com>