Only add compiled groups to the Benchmark
Only add groups as the first ones in the Benchmark if it was
previsously compiled and is loaded in the LinearLoaded.
The same is valid for groups that contain product specific groups, only
add them if they were previously compiled.

If a group Id is not found, it means that it is not applicable to the
product being built.