OCP4/PCI-DSS: Add response for Req-7.1.1
This adds relevant rules to verify that Roles and ClusterRoles are
defined in the cluster. While this may not bring a lot of value given
that we indeed ship a whole lot of roles by default; this makes it so
that the verbiage appears in the report and we have an actual automated
rule.

Signed-off-by: Juan Antonio Osorio Robles <jaosorior@redhat.com>