Clarify local ssg package dependency
The ssg package within this repository is used within controleval.py and
it's also the name of an official package available via pypi. New users
might confuse the local import for the publicly available package, which
will lead to even more confusion since they're obviously two different
packages.

This commit adds a comment to the script to clarify those details for
readers. It also provides additional details as a RuntimeError exception
so that the caller knows what to do if the script fails to import the
local ssg package.