Don't treat 'no_ovirt' and 'ovirt' as packages
Exclude 'no_ovirt' and 'ovirt' during translation from platform to
package. These platforms are about ovirt-engine and ovirt-host packages.

This is a workaround.
At the moment the remediation aspect of these platforms are not well
supported in the project. But with the addition of CPE Applicability
Language, this will be better supported.