General improvements for CTF usage in GHA jobs.
As jobs are expanded to include different branches than the main one,
it's necessary to specify which branch the test is going to run against.
And also expand the usage of finding the fork point so tests are not run
against a newer commit for that specified branch in case the pull
request is not based on latest target branch.