Include the checksum files in the Git Hub release
CMake checksums with `sha512` algorithm were only added in CMake
3.10 and `CPACK_PACKAGE_CHECKSUM` was only added in CPack 3.7.

As these commands are not part of the normal content build process,
I'm not bumping `cmake_minimum_required`. It should be enough to
ensure that the systems doing the release have CMake > 3.10.

Currently, the image used for the releases is a Fedora 34 and it has
CMake 3.20.