Update directory_ownership_var_log_audit
Previous changes that only applied to ol8 were seen that also apply to
rhel8, and the only one else product that uses this rule is rhel9 which
is based on rhel8. Now OVAL passes only if (log_file is set and root
owns its directory) or (log_file is not set and root owns default
directory). Update bash accordingly

Signed-off-by: Edgar Aguilar <edgar.aguilar@oracle.com>