Add a rule for EKS CIS section 5.2
This commit adds one rule for EKS CIS section 5.2. This rule doens't
check anything since it would require access to an AWS IAM endpoint and
knowledge about the deployment itself. This may change in the future,
but for the time being we can at least give the user the Audit and
Remediation sections from the benchmark.