OL8 STIG profile severity adjustments
Update the severity of several rules to be aligned with the
OL8 and RHEL8 STIG profiles.

Additionally, for rules that have a different severity on
OL8 and RHEL8 STIG profiles, adjust the severity on the
OL8 stig profile.

Also remove the 'draft' status from the OL8 stig profile
and update the rule's selection for the OL08-00-010287
requirement.

Signed-off-by: Federico Ramirez <federico.r.ramirez@oracle.com>