On sle15 accounts_authorized_local_users removes accounts it should not
rule accounts_authorized_local_users is active in the sle15 stig.profile
but now entry was ever added to:
linux_os/guide/system/accounts/accounts-restrictions/var_accounts_authorized_local_users_regex.var

for sle15 or sle12.

So create the entried and update sl15 stig profile to use the new entry.
For completeness also update the commented record in the sl12 stig.profile,
but the rule is commented out in the sle12 stig.profile.