Split command line arguments of add_stig_references.py
The option `--unlinked-xccdf` specifies both input file name and
output file name, the script rewrites this file in place. We replace
this option by `input` and `output` arguments which allows us to
specify a different file name for the output of this script or to
keep the input file intact. That doesn't change the purpose of the
script.