Make template check for "n" values
When a kernel config is "n", it is not set in the files.
This changes the template so that when the kernel config value is "n",
the template will check for the absence of the config.

This, unfortunately, creates a mess in the test scenarios, a missing
config entry is a pass for value="n", and fail for value="y".