Stop using Ansible Find module in file_permissions
The 'find' module doesn't allow us to filter by mode, so it returns
a large amount of files, and the 'file' module needs to go through all
of them. This is very slow, and take up to a few hours if we are
checking for permissions in '/usr/lib' for example.

Let's use the 'find' utility via 'command` module to find the files
that need its permissions fixed.

Note: we could even use the 'find' utility to change the file
permissions too, but let's restrict ourselves to the minimum need from
'find' utility.