Improve pam_faillock macro remediation to fix wrong control
Old versions of the pam_faillock remediation created a pam_faillock.so line
with the control "[default=die]", which is not the proper control. The
remediation was already updated for new implementations but the bash
remediation was not updating the control if pam_faillock.so was once
enabled using the old remediation. This patch fix this.