Add tests to check/ensure symlink behavior of OVAl
Add various tests that setup symlinks to files and directories with
correct and incorrect permissions.

As noted in 6a3f401aec68db0510ca9e1006ae29196473951c, the check should
not follow symlinks.

These tests define the expected behavior of the check, witch is to not
follow symlinks nor report symlinks as collected objects.