Fix handling of quotes matching against substrings
The 'pattern match' operation allowed the rule to pass when the
value it is looking for is a substring of the value in /boot/config-*.
The check should actually look for an exact match.

This also unifies the handling of the double quotes in the object, this
way the state for value and variable can only care about matching the
value we are looking for. And here, we need a lazy (or reluctanct) match
so that we don't include the second double quote in the capturing group.