Check for a CNI that supports network policies
Before, configure_network_policies was a manual check that told the user
to make sure they were using a CNI that supported network policies. This
information is actually available through the K8S operator API.

Let's update the rule so that we query the appropriate resource and
filter it so we can check for CNIs that support network policies.

Fixes https://bugzilla.redhat.com/show_bug.cgi?id=2072431