Avoid directly referencing empty object collections
Referencing object definitions that are empty will cause the variable to
result in "no value", and fail the whole test, even when there are
definitions that are not emtpy.

Using sets allows the variable definitnion to reference a "resolved"
collection of objects, the 'set' isolates the problem with empty
object colllections.