SUSE SLE15 create profile for hardening cloud images
SUSE would like to the extent possible to preharden images
that we provide to cloud service providers.

The pcs-hardening.profile was created for this purpose.
It includes all rules from the cis and stig profile that
we determined can safely be applied before to an image
before it is instantiated.

The following rule classes were excluded:
1: anything dealing with passwords, since those require
   user action after the image is booted.
2: anything dealing with graphics. Our cloud images don't include graphics.
3: anything dealing with file partitions. Our cloud images are single
   volume and single partition.
4: package removal. Since we don't know all use cases, we don't want
   to risk creating problems for users by removing software they expect.
5: any rule that does not currently have automated remediation,

The imagined use case is that users would start with a preharned image.
After the VM has been created, users would complete the process using one
of the existing profiles: standard, pci-dss, hipaa, cis or stig.