Fix macro fixtext_audit_rules_unsuccessful_file_modification_o_trunc_write
There can be different positions of the O_TRUNC_WRITE argument
in the syscall parameters in different syscalls, so we can't
hardcode "a2" in the macro text, but we need to allow to specify
the position as a macro paramter. This is similar to how the
audit_rules_unsuccessful_file_modification_o_trunc_write template is
parametrized.