Improve handling templated test scenarios
If a rule is templated we can't assume that all templated scenarios
will be executed. It can have test_config.yaml that blocks some
scenarios and also we have local test scenarios that take precedence
over templated test scenarios. Therefore it can happen that some
templated scenarios won't be executed. So we can't consider the
template fully tested if a rule that uses this template has been
tested. Instead, we need to track templated test scenarios for
each template and attempt to execute as much templated test scenarios
as we can.