Rename "target"
We have different "targets":
- list of rules in rule mode and combined mode
- list of templates in template mode
- list of profiles in profile mode

It can be confusing to use these lists interchageably and
call them "target".

Instead we will separate the list of rules and list of templates
and call them "rule_spec" and "template_spec".

Also fixes a traceback in template mode which happend because
the template mode was introduced to master after the refactoring
PR #8751 started.