OCP4: Add compat rules for checking API server and kubelet TLS certs and keys on OCP releases <= 4.8
The rules that check for the proper API server and kubelet TLS certs and
keys were rewritten recently to support Hypershift style deployments
and use variables in the checks. However, it is not possible to tune
these new rules per-release and OCP releases up to and including 4.8 use
a completely different path for the certs and keys.

Therefore this patch adds four rules that are only valid for the older
releases.