Rework tarball generation
During the test run setup, a tarball with test scenarios and test
data is created and uploaded to the test backend. With this patch,
the tarball will no longer contain test scenarios for the rules that
won't be tested. Also, we will avoid processing test scenarios and
rules twice, we will reuse processed test scenarios that we have
loaded in memory at the moment of the tarball creation.

Fixes: #8765