OCP4: Look at rendered-config-xyz when checking FIPS, consolidate two rules
We used to have two rules that were checking for the cluster being installed
in FIPS mode, one was checking for an MC called 99-master-fips and was
used in the NIST profiles only.

The other rule was looking for both `^[0-9]{2}-worker$` and `^[0-9]{2}-master$`
and was used in PCI-DSS only.

This is all confusing, let's consolidate on one rule that checks the
rendered configs for both master and worker pools. This allow us to use
a single rule regardless of the installation method and also support UPI
installations where the manifests can have a custom name:
    https://docs.openshift.com/container-platform/4.10/installing/install_config/installing-customizing.html

Related: rhbz#2111095