Create rule to cover OL08-00-020262
The rule is file_permissions_lastlog using the template file_owner.
Also fixed call to oval_metadata macro so it shows the content of
FILEPATH variable instead of literally FILEPATH.

Add test for stricter permissions.

And add this rule and the ones about lastlog (group)ownership to OL8
stig profile

Signed-off-by: Edgar Aguilar <edgar.aguilar@oracle.com>