Update the sssd_enable_smartcards rule description
To properly enable smartcards, two pam_sss.so lines are necessary in the
auth group of system-auth file. Following the implementation recommended
by authselect as reference, these two lines have different parameters
and different controls. Also, it is not expected that the smartcard-auth
file has the same content of system-auth file.