Updated accounts_password_pam_retry rule remediation
Due to PAM modules evolution, this rule has a different behaviour based on
the target system version. Since the PAM bash macros were refactored
with flexibility in mind, it is pretty simple to deal with situations like
this. In this specific case, for systems in older versions, the
bash_ensure_pam_module_option acro was directly called instead of
bash_ensure_pam_module_configuration. This way, unecessary code is avoided
in the final remediation while the same standard approach to update PAM files
is preserved.