Regex and descriptions improvemens in PAM macros
The bash_remove_pam_module_option was incremented by a Jinja2
conditional to permit the control parameter to be optional. Althought
recommended to inform the control whenever possible to avoid undesired
changes, there are cases where two or more similar lines with different
controls need to be changed at once. This commit also updates the macros
descriptions expliciting which parameters are optional and mentions type
and group as interchangeable words for the same purpose.