sssd_enable_smartcards: authselect with-smarcard
Use authselect enable-feature with-smartcard to remediate rule
sssd_enable_smartcards.

The Ansible remediation for this rule is breaking the pam stack.
'sudo' stops working after Ansible task is applied.

This commit switches the RHEL8 to use authselect with-smartcard.
I'm aware this moves the rule out of alignment with DISA, but it is
better than having broken 'sudo'.
The alignment needs to be discussed with DISA and fixed later.