Improve test scenarios for accounts_password_pam_retry rule
These test scenarios are valid for sytems where the pwquality.conf file
is present. However, is it possible that a system has the pwquality.conf
file but also have the pam_pwquality.so parameters defined in PAM files.
The scripts are now ensuring the parameter is not present in PAM files.