Update ansible_ensure_pam_module_line macro
The macro description was updated according to the Ansible remediation,
which is slightly different from the respective Bash remediation due to
technical reasons. In Ansible, the safeguard to modify the control only
for the first match in rare cases was not possible with considerable
complexity. The tests proved this safeguard is not essential for the
mapped cases. Therefore, the description was updated to make clear the
macro behavior and warn about possible atypical cases. This commit also
adjust the Ansible condition for cases where more lines are matched.