Create macro for authselect backup - Ansible
Many Ansible tasks related to authselect eventually needs to run the
"authselect apply-changes" command to apply changes and automatically
create a backup. Instead of repeating this command, it was created a
macro to ensure it is properly executed. This macro also provides a
parameter which permits changing the default backup name. If not
informed, the authselect standard is used.