Enhance the bash_ensure_pam_module_line macro capabilities
In Ansible, it is accepted the special value "BOF" for "insertafter"
parameter from the "lineinfile" module, which instructs the module to insert
a new line at the beginning of the file. In bash the "sed" command is
used. To achieve the same effect, the sed command needs to slightly
changed. So, the macro ensures the necessary change if "BOF" is
informed in the "after_match" parameter.