Update OVAL assessment for sssd_enable_smartcards
The system-auth and smartcard-auth PAM files have different content in a
proper implementation with smartcard enabled. The OVAL assessment was
properly updated to differentiate the requirements of these two files.
The check conditionals were reviewed and updated to avoid false positives
where the file exists but is not compliant. The service check was removed
since it is not expected in this rule. Therefore, unecessary checks were removed.