Relaxed the chrony user check
I have verified that the '-u chrony' in OPTIONS is not needed
on RHEL7-9, nor on Fedora.
The chrony user is the one that is used to run the service.
As a result, I have updated the check, remediations and description.

The behavior stays the same for the other systems.