Don't pass sssd rules when sssd.conf is absent
Some sssd rules check if sssd.conf exists and pass the rule if the file
is missing, but other rules don't check for it, like rule
sssd_run_as_sssd_user, sssd_enable_smartcards and
sssd_enable_pam_services.

The assumption is that the SSSD service is not in use when the config
file is missing. But the passing can mislead one into thinking the file
is configured.

This also makes the behaviour among sssd rules consistent.