Improve bash_ensure_pam_module_line macro
This macro was previously updated to align to the Ansible equivalent in
cases where multiple matches were found, in order to only include a line
after the last match. However, in cases where the informed regex have no
match, many lines would be included. This patch ensures the same
behavior of the Ansible equivalent when there is no match.