Ignore invalid or non existing include objects
Let's not fail the task when the find doesn't find the include object.
When the include is a glob expression that doesn't evaluate to any file
the glob itself is used in find command.

The Bash remediation prints a message for each include that is not a
file is not a directory or doesn't exist.