Include warning about the pam_securetty.so PAM module
The no_direct_root_logins rule checks and remediate the /etc/securetty file,
which is used by the pam_securetty.so module. Currently, this module is no
longer enabled by default in many distros. It was included a warning
message to make the user aware of this, since the /etc/securetty file is
useless without the pam_securetty.so module properly enabled.