Mask password field in accounts_password_all_shadowed_sha512
There are filters in this rule to exclude compliant or irrelevant objects.
However, a non compliant object should have the password field hidden in
reports. Minor changes were needed in the OVAL test, as the mask parameter
has no effect when used to filter objects. So, in order not to make major
changes to the OVAL logic, a state was included in the test.