Fix issues with bash remediation
The bash remediation was not prepared for a situation where more than
one directory is defined, for example, in faillock.conf and in PAM
files. In situations like this, the remediation would fail. Also, the
remediation was not considering an existing configuration for SELinux
context. For example, if the directory was already defined via semanage,
an error is returned if tried to insert it again. In this case, the
SELinux context should be updated with appropriate command.