Fix check_existence enumeration
The account_password_selinux_faillock_dir rule intended to fix the
SELinux fcontext in faillock data dirs when they are defined. However,
there is no necessary action if the faillock dirs are not defined in
the system. If dirs are not set, there is no fcontext issue and the rule
should pass.