Fix missing OCIL text and 800-53 references
In #9325 we discovered that the
`table-${PRODUCT}-${STIG_PROFILE}-testinfo.html` is missing contents in
`Check Text (OCIL Check)` and `800-53 Refs` columns.

However, it isn't caused by #9325, these data were empty even before,
the reasons are:
- the template doesn't expect OCIL as a separate file, it expects
  OCIL data embedded in XCCDF as it used to be in past
- the template used a wrong XML namespace for CCI list so the
  CCI list couldn't be parsed by the template.

This commit tries to solve these 2 problems. However, there are
offensive `<sub>` elements in the HTML output which need still to be
resolved on the level of `ocil-unlinked.xml` generator (it isn't a
problem in this template).