Fix namespace of select elements
In #9345 we have started to create elements with explicit XCCDF 1.1
namespace. However, meanwhile #9332 got merged and that adds a new
block that creates a `select` element, however, this wasn't updated
to create elements with explicit XCCDF 1.1 namespace. After merging
both #9345 and #9332, this omission caused that the namespace
of the `select` element was invalid and the build got groken.