Conditional for Ansible remediation on RHEL7
The Ansible remediation for the accounts_password_set_max_life_existing
rule was recently updated to use the user module with the
password_expire_max parameter. However, this parameter was introduced
on Ansible 2.11 while RHEL7 uses Ansible 2.9. Therefore, the former
remediation approach was kept for RHEL7 via a Jinja2 conditional.