Include informative message regaring profile selection
The authselect profile selection command will fail if the PAM settings
were previously modified without authselect. In some cases, it is safe
to include the force option in the command. This will overwrite the PAM
settings according to the selected profile. It is important the user is
aware of this behaviour when using the force option.