rewrite test_machine_only_rules test
do not check content of groups in the datastream, platforms should not appear on multiple places after the change in the previous commit
rather check for presence of conditionals in remediations
do not check for presence of "platform" tag in the "rule" definition because we are no longer copying the platform tag from parent groups. We use XCCDF parsing mechanism to correctly make rule applicable or not.