Remove superflous check of rule ID consistency
The `name` attribute of `check-content-ref` element is generated by
build_shorthand.py and its value is set there to rule's ID, so we don't
need to verify that its set to rule's ID.

See:
https://github.com/ComplianceAsCode/content/blob/54f40fc7d4ea1b9bbc76a575ccb7ff91fc1c109e/ssg/build_yaml.py#L1685
and
https://github.com/ComplianceAsCode/content/blob/54f40fc7d4ea1b9bbc76a575ccb7ff91fc1c109e/ssg/build_yaml.py#L1694

The check code is a leftower from past when the shorthand file used to
be generated by XSLTs.

Don't get confused, the `checks` variable in the removed code actually
contains the `check` element and the `check` variable is actually
a `check-content-ref` element.